backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
google-genai>=1.0.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
